+++ metrics_utility/anonymized_rollups/execution_environments_anonymized_rollup.py
